Added handling for the callsets/{id}/calls call
The is a first attempt at this new endpoint. Still needs fleshed out a
little more, but I wanted to commit where I got to with it. It also
hints at a direction we may move in for the response structure classes
as until now we didn't have a clean way to hanle master/detail type
responses.