Forget about using inheritance for Pagination
There is no clean way to do this. For now just create concrete
implementations of the few classes that need the different Pagination
style (and so different metadata object) and revisit this later to see
if there's a cleaner way to model the data. I have been wanting to do
this anyway.